Do you guys want to design the site like this?
Left Column:

The left column would push channel elements like a stack data structure whenever a user joins a conversation. Inside each pushed channel, a circle with a number in the middle of it will show how many new messages have been sent in the channel while browsing other ones. Also if the user wants to leave a channel, it will be pushed out of the stack and removed from the left column. When the user clicks on one of the channels, then the browser will open the channel in the middle column.

Middle Column:

Middle column will be endless with user created channels! This is the community of channels so a user can browse through and pick which channel to be apart of. When the user clicks on the channel it will use a drop down box to display the channels conversation. 

Right Column:

When a user is in a channel, the right column will display the details of the current channel created by the user who created the channel. Below that if we integrate pr